+++ extensions/quick-access/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Update file preview] - 2022-05-25
 
 - Update file preview
+- Update "Show in Finder" action shortcuts
+- Add action "Open Extension Preferences"
 
 ## [Update file preview] - 2022-04-30